Update: context and refactor
